+++ src/main/java/ladder/controller/LadderGameController.java
@@ -6,7 +6,7 @@ import ladder.domain.ladder.Block;
 import ladder.domain.ladder.Ladder;
 import ladder.domain.ladder.Line;
 import ladder.domain.player.Players;
-import ladder.exception.CustomException;
+import ladder.common.CustomException;
 import ladder.view.InputView;
 import ladder.view.OutputView;
