+++ src/main/java/ladder/domain/player/exception/PlayerNameLengthException.java
@@ -1,4 +1,6 @@
-package ladder.exception;
+package ladder.domain.player.exception;
+
+import ladder.common.CustomException;
 
 public class PlayerNameLengthException extends CustomException {
