+++ src/main/java/ladder/domain/player/exception/PlayerNumberException.java
@@ -1,4 +1,6 @@
-package ladder.exception;
+package ladder.domain.player.exception;
+
+import ladder.common.CustomException;
 
 public class PlayerNumberException extends CustomException {
